Remove test file
By this file I wanted to check what modification date the file has
if we pull it from the repository. That helps me with decide
to use Makefile to not to install node modules and vendor
every time when want to has them up to date but only if they
not exist or they are older than dependency definitions files.
It works right I wanted :D