Replace read models of kart and track to just models
The previous reason to create them, was to separate the module and
the application, and couple them only with events. This is great
solution for modules that have some logic inside, and separating
read and write model make sens. Right there the only sense to keep
read models of CRUD's was to separate them from framework.
In laravel when you use the Eloquent it doesn't make any sens,
because they are coupled at all because of extending model.
Using the CRUD model only, will decrease complexity.