Replace string controller definitions to callables
Laravel allow us to define default namespace for specific
routing configuraion, and then to define controllers and
methods as a short strings.
This is not the best solution. Better is to use the callables
using concrete reference to class. This is more usefull
for example if you want to navigate to the specific
controller using the IDE functionality. It's also help us
avoiding typos in configuration.